requirements.txt
# Framework web
Flask==3.0.0
Flask-CORS==4.0.0

# Base de datos
#psycopg2-binary==2.9.9        # PostgreSQL driver, descomentar cambiar para docker
psycopg[binary]
SQLAlchemy==2.0.23            # ORM para Postgres
pymongo==4.6.1                # MongoDB driver

# Autenticación
PyJWT==2.8.0                  # JSON Web Tokens
bcrypt==4.1.2                 # Password hashing

# Validación
marshmallow==3.20.1           # Serialización y validación

# Utilidades
python-dotenv==1.0.0          # Variables de entorno
pytz==2023.3                  # Zonas horarias

# Logging
python-json-logger==2.0.7     # Logs estructurados

# Testing (opcional para desarrollo)
pytest==7.4.3
pytest-flask==1.3.0